# use Dict to store counts in Composition