--- a/src/composition.jl
+++ b/src/composition.jl
@@ -6,8 +6,14 @@
 # This file is a part of BioJulia.
 # License is MIT: https://github.com/BioJulia/BioSequences.jl/blob/master/LICENSE.md
 
+"""
+Sequence composition.
+
+This is a subtype of `Associative{T,Int}`, and the `getindex` method returns the
+number of occurrences of a symbol or a k-mer.
+"""
 struct Composition{T} <: Associative{T,Int}
-    counts::Vector{Int}
+    counts::Dict{T,Int}
 end
 
 function Composition(seq::BioSequence{A}) where {A<:Union{DNAAlphabet,RNAAlphabet}}
@@ -15,7 +21,7 @@ function Composition(seq::BioSequence{A}) where {A<:Union{DNAAlphabet,RNAAlphabe
     @inbounds for x in seq
         counts[reinterpret(UInt8, x) + 1] += 1
     end
-    return Composition{eltype(A)}(counts)
+    return Composition{eltype(A)}(count_array2dict(counts, alphabet(A)))
 end
 
 function Composition(seq::ReferenceSequence)
@@ -23,18 +29,25 @@ function Composition(seq::ReferenceSequence)
     @inbounds for x in seq
         counts[reinterpret(UInt8, x) + 1] += 1
     end
-    return Composition{DNA}(counts)
+    return Composition{DNA}(count_array2dict(counts, ACGTN))
 end
 
-function Composition(kmer::Kmer{T,k}) where {T,k}
-    counts = zeros(Int, 16)
-    @inbounds begin
-        counts[Int(DNA_A)+1] = count_a(kmer)
-        counts[Int(DNA_C)+1] = count_c(kmer)
-        counts[Int(DNA_G)+1] = count_g(kmer)
-        counts[Int(DNA_T)+1] = count_t(kmer)  # U when T == RNA
-    end
-    return Composition{T}(counts)
+function Composition(kmer::DNAKmer)
+    counts = Dict{DNA,Int}()
+    counts[DNA_A] = count_a(kmer)
+    counts[DNA_C] = count_c(kmer)
+    counts[DNA_G] = count_g(kmer)
+    counts[DNA_T] = count_t(kmer)
+    return Composition(counts)
+end
+
+function Composition(kmer::RNAKmer)
+    counts = Dict{RNA,Int}()
+    counts[RNA_A] = count_a(kmer)
+    counts[RNA_C] = count_c(kmer)
+    counts[RNA_G] = count_g(kmer)
+    counts[RNA_U] = count_t(kmer)
+    return Composition(counts)
 end
 
 function Composition(seq::AminoAcidSequence)
@@ -42,13 +55,28 @@ function Composition(seq::AminoAcidSequence)
     @inbounds for x in seq
         counts[reinterpret(UInt8, x) + 1] += 1
     end
-    return Composition{AminoAcid}(counts)
+    return Composition{AminoAcid}(count_array2dict(counts, alphabet(AminoAcid)))
 end
 
 function Composition(iter::EachKmerIterator{T}) where {T<:Kmer}
-    counts = zeros(Int, 4^kmersize(T))
-    for (_, x) in iter
-        counts[convert(UInt64, x) + 1] += 1
+    counts = Dict{T,Int}()
+    if kmersize(T) ≤ 8
+        # This is faster for short k-mers.
+        counts′ = zeros(Int, 4^kmersize(T))
+        for (_, x) in iter
+            @inbounds counts′[reinterpret(Int, x)+1] += 1
+        end
+        for x in 1:endof(counts′)
+            @inbounds c = counts′[x]
+            if c > 0
+                counts[reinterpret(T, x-1)] = c
+            end
+        end
+    else
+        for (_, x) in iter
+            get!(counts, x, 0)
+            counts[x] += 1
+        end
     end
     return Composition{T}(counts)
 end
@@ -71,29 +99,23 @@ function Base.length(comp::Composition)
 end
 
 function Base.start(comp::Composition)
-    return UInt64(0)
+    return start(comp.counts)
 end
 
-function Base.done(comp::Composition, i)
-    return i ≥ length(comp.counts)
+function Base.done(comp::Composition, s)
+    return done(comp.counts, s)
 end
 
-function Base.next(comp::Composition, i)
-    key = convert(keytype(comp), i)
-    count = comp.counts[i + 1]
-    return (key => count), i + 1
+function Base.next(comp::Composition, s)
+    return next(comp.counts, s)
 end
 
 function Base.getindex(comp::Composition{T}, x) where {T}
-    i = convert(UInt64, convert(T, x))
-    if !(0 ≤ i < endof(comp.counts))
-        throw(KeyError(x))
-    end
-    return comp.counts[i+1]
+    return get(comp.counts, convert(T, x), 0)
 end
 
-function Base.copy(comp::Composition{T}) where {T}
-    return Composition{T}(copy(comp.counts))
+function Base.copy(comp::Composition)
+    return Composition(copy(comp.counts))
 end
 
 function Base.merge(comp::Composition{T}, other::Composition{T}) where {T}
@@ -101,9 +123,8 @@ function Base.merge(comp::Composition{T}, other::Composition{T}) where {T}
 end
 
 function Base.merge!(comp::Composition{T}, other::Composition{T}) where {T}
-    @assert length(comp.counts) == length(other.counts)
-    for i in 1:endof(comp.counts)
-        comp.counts[i] += other.counts[i]
+    for (x, c) in other
+        comp.counts[x] = comp[x] + c
     end
     return comp
 end
@@ -118,4 +139,16 @@ function Base.summary(::Composition{T}) where {T}
     else
         return string(T, " Composition")
     end
+end
+
+function count_array2dict(counts, alphabet)
+    counts′ = Dict{eltype(alphabet),Int}()
+    sizehint!(counts′, countnz(counts))
+    for x in alphabet
+        @inbounds c = counts[convert(Int64,x)+1]
+        if c > 0
+            counts′[x] = c
+        end
+    end
+    return counts′
 end