This is a little bit slower than using Vector but it is able to store
counts of long k-mers with far less memory.