@Ward9250 2017-08-07
Could we have both versions? Unified by an API and abstract type, and then the best one for the use case applied?