@bicycle1885 2017-08-07
We can, but I don't think it is a good idea since it makes the code more complicated. The new data structure I propose is still fast in most cases and the counting algorithm is not changed. So, the performance of `composition` does not change for `BioSequence`s:

master:
```
julia> @benchmark composition($seq)
BenchmarkTools.Trial:
  memory estimate:  224 bytes
  allocs estimate:  2
  --------------
  minimum time:     27.222 μs (0.00% GC)
  median time:      29.200 μs (0.00% GC)
  mean time:        30.333 μs (0.00% GC)
  maximum time:     212.823 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1

```

this branch:
```
julia> @benchmark composition($seq)
BenchmarkTools.Trial:
  memory estimate:  800 bytes
  allocs estimate:  10
  --------------
  minimum time:     27.721 μs (0.00% GC)
  median time:      28.023 μs (0.00% GC)
  mean time:        30.111 μs (0.00% GC)
  maximum time:     137.682 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1

```

The difference gets large for `Kmer`s but using `composition` for a `Kmer` value is not best in terms of performance because it does some allocation. Using `count_{a,c,g,t}` is the best choice in that case.

master:
```
julia> @benchmark composition(kmer"ACGTAGTA")
BenchmarkTools.Trial:
  memory estimate:  224 bytes
  allocs estimate:  2
  --------------
  minimum time:     42.198 ns (0.00% GC)
  median time:      46.022 ns (0.00% GC)
  mean time:        54.982 ns (11.49% GC)
  maximum time:     1.431 μs (86.91% GC)
  --------------
  samples:          10000
  evals/sample:     987

julia> @benchmark BioSequences.count_a(kmer"ACGTAGTA")
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     1.896 ns (0.00% GC)
  median time:      2.050 ns (0.00% GC)
  mean time:        2.210 ns (0.00% GC)
  maximum time:     35.214 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000

```

this branch:
```
julia> @benchmark composition(kmer"ACGTAGTA")
BenchmarkTools.Trial:
  memory estimate:  592 bytes
  allocs estimate:  9
  --------------
  minimum time:     255.330 ns (0.00% GC)
  median time:      271.447 ns (0.00% GC)
  mean time:        322.011 ns (9.04% GC)
  maximum time:     6.661 μs (85.65% GC)
  --------------
  samples:          10000
  evals/sample:     352

```