@Ward9250 2017-08-07
That sounds fair enough. Perhaps I should have a look at using Composition with the fast pairwise counting mechanism I built.