@Ward9250 2017-08-08
If I dict is used for the composition, is there any point to the wrapping associative type anymore? It original purpose seems to be to act like a dict, but using a vector underneath, now a dict is used underneath, is there a point to the wrapping type anymore?