@bicycle1885 2017-08-09
Yeah, that makes sense, but the semantic is significantly different from dicts. For example, `getindex` is overloaded to return zero for non-exist elements, which is similar to `DefautDict` of DataStructures.jl. Also, `merge` adds two counts instead of overwriting the former with the latter. `Composition` is convertible to `Dict`, too.